+++ web.rb
@@ -13,10 +13,9 @@ require_relative 'lib/mu_search/index_builder.rb'
 require_relative 'lib/mu_search/search_index.rb'
 require_relative 'lib/mu_search/index_manager.rb'
 require_relative 'lib/mu_search/elastic.rb'
+require_relative 'lib/mu_search/tika.rb'
 require_relative 'framework/elastic_query_builder.rb'
-require_relative 'framework/tika.rb'
 require_relative 'framework/jsonapi.rb'
-
 
 ##
 # WEBrick setup
@@ -44,7 +43,7 @@ end
 ##
 # Setup index manager based on configuration
 ##
-def setup_index_manager(elasticsearch, tika, sparql_connection_pool, config)
+def setup_index_manager(elasticsearch, config)
   search_configuration = config.select do |key|
     [:type_definitions, :default_index_settings,
      :persist_indexes, :eager_indexing_groups, :number_of_threads,
@@ -55,15 +54,13 @@ def setup_index_manager(elasticsearch, tika, sparql_connection_pool, config)
   MuSearch::IndexManager.new(
     logger: Mu::log,
     elasticsearch: elasticsearch,
-    tika: tika,
-    sparql_connection_pool: sparql_connection_pool,
     search_configuration: search_configuration)
 end
 
 ##
 # Setup delta handling based on configuration
 ##
-def setup_delta_handling(index_manager, elasticsearch, tika, sparql_connection_pool, config)
+def setup_delta_handling(index_manager, elasticsearch, config)
   if config[:automatic_index_updates]
     search_configuration = config.select do |key|
       [:type_definitions, :number_of_threads, :update_wait_interval_minutes,
@@ -73,8 +70,6 @@ def setup_delta_handling(index_manager, elasticsearch, tika, sparql_connection_p
       logger: Mu::log,
       index_manager: index_manager,
       elasticsearch: elasticsearch,
-      tika: tika,
-      sparql_connection_pool: sparql_connection_pool,
       search_configuration: search_configuration)
   else
     search_configuration = config.select do |key|
@@ -88,7 +83,6 @@ def setup_delta_handling(index_manager, elasticsearch, tika, sparql_connection_p
 
   delta_handler = MuSearch::DeltaHandler.new(
     logger: Mu::log,
-    sparql_connection_pool: sparql_connection_pool,
     update_handler: handler,
     search_configuration: { type_definitions: config[:type_definitions] })
   delta_handler
@@ -104,37 +98,27 @@ configure do
   configuration = MuSearch::ConfigParser.parse('/config/config.json')
   set configuration
 
-  tika = Tika.new(
-    host: 'tika',
-    port: 9998,
-    logger: Mu::log
-  )
+  connection_pool_size = configuration[:connection_pool_size]
+  MuSearch::Tika::ConnectionPool.setup(size: connection_pool_size)
 
-  elasticsearch = MuSearch::Elastic.new(
-    host: 'elasticsearch',
-    port: 9200,
-    logger: Mu::log
-  )
+  elasticsearch = MuSearch::Elastic.new(size: connection_pool_size)
   set :elasticsearch, elasticsearch
 
-  sparql_connection_pool = MuSearch::SPARQL::ConnectionPool.new(
-    number_of_threads: configuration[:number_of_threads],
-    logger: Mu::log
-  )
+  MuSearch::SPARQL::ConnectionPool.setup(size: connection_pool_size)
 
   until elasticsearch.up?
     Mu::log.info("SETUP") { "...waiting for elasticsearch..." }
     sleep 1
   end
 
-  until sparql_connection_pool.up?
+  until MuSearch::SPARQL::ConnectionPool.up?
     Mu::log.info("SETUP") { "...waiting for SPARQL endpoint..." }
     sleep 1
   end
 
-  index_manager = setup_index_manager elasticsearch, tika, sparql_connection_pool, configuration
+  index_manager = setup_index_manager elasticsearch, configuration
   set :index_manager, index_manager
-  delta_handler = setup_delta_handling index_manager, elasticsearch, tika, sparql_connection_pool, configuration
+  delta_handler = setup_delta_handling index_manager, elasticsearch, configuration
   set :delta_handler, delta_handler
 end